includes for all; new logo; background pic; fixed nav; created DTO package; DONE with initial advQuery RnD
