+++ src/main/webapp/WEB-INF/include/bodyDesign.jsp
@@ -1,0 +1,5 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+
+<%@ taglib prefix = "c" uri = "http://java.sun.com/jsp/jstl/core" %>
+
+<body class="" style="background-image: url('/img/pexels_autumnTrees.jpg'); background-repeat: no-repeat; background-attachment: fixed; background-size: cover;">
