+++ src/main/webapp/WEB-INF/include/footerbuffer.jsp
@@ -2,6 +2,4 @@
 
 <%@ taglib prefix = "c" uri = "http://java.sun.com/jsp/jstl/core" %>
 
-<div id=footerbuffer class="container-fluid" style="height: 4rem;">
-	
-</div>
+<div id=footerbuffer class="container-fluid" style="height: 4rem;"></div>
