+++ src/main/webapp/WEB-INF/include/pageLayoutBottom.jsp
@@ -1,0 +1,15 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+
+<%@ taglib prefix = "c" uri = "http://java.sun.com/jsp/jstl/core" %>
+
+			</div> <!-- end col -->
+			<div class="col py-0">
+			
+			</div> <!-- end col -->
+		</div> <!-- end row -->
+	</div><!-- end main -->
+
+	<%-- <jsp:include page="/WEB-INF/include/footerbuffer.jsp"/> --%>
+ 	<jsp:include page="/WEB-INF/include/footer.jsp"/>
+ </body>
+</html>
