+++ src/main/webapp/WEB-INF/include/pageLayoutTop.jsp
@@ -1,0 +1,10 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+
+<%@ taglib prefix = "c" uri = "http://java.sun.com/jsp/jstl/core" %>
+
+<!-- <div id=main class="container-fluid" > -->
+<div id=main class="container-fluid" >
+		<div class="row" style="height: 100vh" >
+			<div class="col py-0">
+			</div> <!-- end col -->
+			<div class="col-sm-8 bg-light py-0"  >
